chore: re-organize code of SplitChunksPlugin
