+++ crates/rspack_plugin_split_chunks/src/max_size_queue_item.rs
@@ -1,0 +1,9 @@
+use crate::cache_group_source::SplitChunkSizes;
+
+#[derive(Debug)]
+pub struct MaxSizeQueueItem {
+  pub min_size: SplitChunkSizes,
+  pub max_async_size: SplitChunkSizes,
+  pub max_initial_size: SplitChunkSizes,
+  pub keys: Vec<String>,
+}
